eds-symptom-discovery/requirements.txt
# EDS Symptom Discovery Tool - Python Dependencies

# Data Collection
praw==7.7.1                    # Reddit API wrapper
requests==2.31.0               # HTTP library for PubMed API
beautifulsoup4==4.12.2         # HTML parsing (if needed for criteria)

# Data Processing
pandas==2.1.4                  # Data manipulation
numpy==1.26.2                  # Numerical computing

# NLP & Text Analysis
nltk==3.8.1                    # Natural language toolkit
spacy==3.7.2                   # Advanced NLP (optional)

# Analysis & Mining
scikit-learn==1.3.2            # Machine learning utilities
scipy==1.11.4                  # Scientific computing

# Visualization
matplotlib==3.8.2              # Plotting
seaborn==0.13.0                # Statistical visualization

# Jupyter
jupyter==1.0.0                 # Notebook environment
ipykernel==6.27.1              # Jupyter kernel
tqdm==4.66.1                   # Progress bars

# Utilities
python-dotenv==1.0.0           # Environment variables

# LLM/AI
openai==1.54.3                 # OpenAI API for symptom extraction
